lambda-python-performance/pyproject.toml
[tool.poetry]
name = "lambda-python-performance"
version = "0.1.0"
description = ""
authors = ["John Franey <1728528+johnfraney@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "lambda_python_performance"}]

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-aws-lambda-python-alpha = "^2.76.0a0"
httpx = "^0.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
aws-cdk-lib = "2.76.0"
black = "^23.3.0"
ruff = "^0.0.262"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
